pyproject.toml
[project]
name = "hydra-c2"
version = "0.1.0"
description = "HYDRA-C2: Hybrid Universal Dynamic Reconnaissance Architecture - Open-source Multi-Domain C2 System"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "HYDRA-C2 Team" }]

dependencies = [
    # === Layer 4: Persistence ===
    "psycopg[binary]>=3.1",      # PostgreSQL async driver
    "asyncpg>=0.29",               # PostgreSQL async driver for SQLAlchemy
    "geoalchemy2>=0.14",          # PostGIS ORM support
    "neo4j>=5.20",                # Neo4j Python driver
    "sqlalchemy>=2.0",            # ORM / query builder

    # === Layer 3: Data Ingestion ===
    "paho-mqtt>=2.1",             # MQTT client (Mosquitto)
    "lxml>=5.0",                  # CoT XML parsing
    "xmltodict>=0.13",            # XML → dict conversion

    # === Layer 5: Analytics ===
    "pandas>=2.2",
    "geopandas>=0.14",
    "networkx>=3.3",              # Graph analytics
    "scikit-learn>=1.5",          # ML (Isolation Forest, DBSCAN)
    "shapely>=2.0",               # Geometric operations
    "folium>=0.16",               # Interactive maps

    # === Layer 6: Visualization / API ===
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "websockets>=12.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.0",
    "httpx>=0.27",

    # === Cross-cutting ===
    "python-dotenv>=1.0",
    "structlog>=24.1",            # Structured logging
    "cryptography>=42.0",         # AES-256, TLS
    "pyyaml>=6.0",
    "typer>=0.12",                # CLI interface
    "rich>=13.7",                 # Terminal UI
]

[project.optional-dependencies]
sdr = [
    "pyrtlsdr>=0.3",             # RTL-SDR
    "pyais>=2.6",                 # AIS decoding
]
spark = [
    "pyspark>=3.5",
    "delta-spark>=3.1",
]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.scripts]
hydra = "hydra_c2.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/hydra_c2"]
